requirements.txt
beautifulsoup4>=4.12.3
faiss-cpu>=1.8.0
gradio>=4.44.0
langdetect>=1.0.9
lxml>=5.2.2
numpy>=1.26.4
requests>=2.32.3
sentence-transformers>=3.0.1
tqdm>=4.66.4
